Fixed an error in check for existing smilies
Also, addslashes() is already applied to the input smilie before calling, so I'm not sure whether it's correct to perform addslashes() again inside the function. I haven't checked whether it actually has to be applied twice.

However, the array thing was definitely wrong.

I know this hasn't been touched for 5 years, but there are still vBulletin 3.7 forums out there, and if I'm using this I can't be the only one.